backend/requirements.txt
flask
psycopg[binary]
python-dotenv
flask-cors